// src/app/interfaces/record.ts
export interface Record {
    id: number;
    br: string;
    korisnikUsluge: string;
    mestoRada: string;
    rukovaoc: string;
    radnoVremeOd: Date;
    radnoVremeDo: Date;
    datum: Date;
    vrstaMasine: string;
    sifra: string;
    stanjeMotocasovnikaOd: string;
    stanjeMotocasovnikaDo: string;
}